+++ client/src/components/FindUserForm.tsx
@@ -18,7 +18,7 @@ function FindUserFormI({ setUserList }: FindUserFormI) {
   const onSubmit = (e: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
     e.preventDefault();
 
-    // if (filter.num===0)
+    filter.num = filter.num?.replace(/-/g, "");
     //validation
     const { error } = joiSchema.validate(filter);
     if (error) {
@@ -37,6 +37,19 @@ function FindUserFormI({ setUserList }: FindUserFormI) {
     getUsers().then((users) => setUserList(users));
   };
 
+  const numOnKeyUp = (e: React.KeyboardEvent<HTMLInputElement>) => {
+    let { value } = e.currentTarget;
+
+    // mask. "220055" => "22-00-55"
+    value = value.replace(/-/g, "");
+    value = value.replace(/(\d{2})(\d{1})/, "$1-$2");
+    value = value.replace(/(\d{2})-(\d{2})(\d{1})/, "$1-$2-$3");
+    if (value.length > 8) {
+      value = value.slice(0, 8);
+    }
+
+    setFilter({ ...filter, num: value });
+  };
   return (
     <div className="search-form">
       <h2>search</h2>
@@ -50,11 +63,12 @@ function FindUserFormI({ setUserList }: FindUserFormI) {
           onChange={(e) => setFilter({ ...filter, email: e.target.value })}
         />
         <UI.Input
-          type="number"
+          type="text"
           name="user-number"
           id="user-number"
           placeholder="number"
           value={filter.num}
+          onKeyUp={numOnKeyUp}
           onChange={(e) => setFilter({ ...filter, num: e.target.value })}
         />
         <UI.Button onClick={(e) => onSubmit(e)}>search</UI.Button>