+++ CHANGELOG.md
@@ -7,3 +7,9 @@
 ## 0.8.0
 
 - complete frontend
+
+## 1.0.0
+
+- added mask to 3205.client/ src/ components/ FindUserForm.tsx
+- complete bash run
+- added "TODO.md " with a technical assignment inside